Sort sprites by layer
Added in a call to sort sprites by layer so that higher layered sprites
are drawn later.

It's a fairly inefficient hack since it sorts the sprite list every
frame. If we were using Pygame-native sprites, which have this
functionality built in, this would probably go faster... so we might
want to find a way to do that even with the front end decoupled from the
back end.