Added a priority attribute to sprites
This is so sprites that tend to affect other sprites can be called first
during the turn, so that those effects can propagate to other sprites.

It does mean doing a sort every time, so it can definitely be optimized,
but for now it works.

The chained filtering can be optimized too...